deploy desde heroku modifique persona sin dto y interface

+++ src/main/java/com/apportfolio/mmm/Repository/PersonaRepo.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PersonaRepo extends JpaRepository<Persona, Long> {
+public interface PersonaRepo extends JpaRepository<Persona, Integer> {
 
 }
